requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-phonenumber-field==6.3.0
phonenumbers==8.12.52
Pillow==9.2.0
sqlparse==0.4.2
